backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pandas
sqlalchemy
psycopg2-binary
openai
pytz
geopandas
shapely
fiona
pyproj
rtree
chromadb